+++ zerver/lib/narrow.py
@@ -1,16 +1,13 @@
-import os
 import re
 from dataclasses import dataclass
 from typing import (
     Any,
     Callable,
-    Collection,
     Dict,
     Generic,
     Iterable,
     List,
     Optional,
-    Protocol,
     Sequence,
     Set,
     Tuple,
@@ -48,7 +45,7 @@ from typing_extensions import TypeAlias, override
 from zerver.lib.addressee import get_user_profiles, get_user_profiles_by_ids
 from zerver.lib.exceptions import ErrorCode, JsonableError
 from zerver.lib.message import get_first_visible_message_id
-from zerver.lib.narrow_helpers import NarrowTerm
+from zerver.lib.narrow_predicate import channel_operators, channels_operators
 from zerver.lib.recipient_users import recipient_for_user_profiles
 from zerver.lib.sqlalchemy_utils import get_sqlalchemy_connection
 from zerver.lib.streams import (
@@ -58,7 +55,6 @@ from zerver.lib.streams import (
     get_stream_by_narrow_operand_access_unchecked,
     get_web_public_streams_queryset,
 )
-from zerver.lib.topic import RESOLVED_TOPIC_PREFIX, get_topic_from_message_info
 from zerver.lib.topic_sqlalchemy import (
     get_resolved_topic_condition_sa,
     topic_column_sa,
@@ -80,34 +76,6 @@ from zerver.models.users import (
     get_user_by_id_in_realm_including_cross_realm,
     get_user_including_cross_realm,
 )
-
-stop_words_list: Optional[List[str]] = None
-
-
-def read_stop_words() -> List[str]:
-    global stop_words_list
-    if stop_words_list is None:
-        file_path = os.path.join(
-            settings.DEPLOY_ROOT, "puppet/zulip/files/postgresql/zulip_english.stop"
-        )
-        with open(file_path) as f:
-            stop_words_list = f.read().splitlines()
-
-    return stop_words_list
-
-
-# "stream" is a legacy alias for "channel"
-channel_operators: List[str] = ["channel", "stream"]
-# "streams" is a legacy alias for "channels"
-channels_operators: List[str] = ["channels", "streams"]
-
-
-def check_narrow_for_events(narrow: Collection[NarrowTerm]) -> None:
-    supported_operators = [*channel_operators, "topic", "sender", "is"]
-    for narrow_term in narrow:
-        operator = narrow_term.operator
-        if operator not in supported_operators:
-            raise JsonableError(_("Operator {operator} not supported.").format(operator=operator))
 
 
 def is_spectator_compatible(narrow: Iterable[Dict[str, Any]]) -> bool:
@@ -144,64 +112,6 @@ def is_web_public_narrow(narrow: Optional[Iterable[Dict[str, Any]]]) -> bool:
         and term["negated"] is False
         for term in narrow
     )
-
-
-class NarrowPredicate(Protocol):
-    def __call__(self, *, message: Dict[str, Any], flags: List[str]) -> bool: ...
-
-
-def build_narrow_predicate(
-    narrow: Collection[NarrowTerm],
-) -> NarrowPredicate:
-    """Changes to this function should come with corresponding changes to
-    NarrowLibraryTest."""
-    check_narrow_for_events(narrow)
-
-    def narrow_predicate(*, message: Dict[str, Any], flags: List[str]) -> bool:
-        def satisfies_operator(*, operator: str, operand: str) -> bool:
-            if operator in channel_operators:
-                if message["type"] != "stream":
-                    return False
-                if operand.lower() != message["display_recipient"].lower():
-                    return False
-            elif operator == "topic":
-                if message["type"] != "stream":
-                    return False
-                topic_name = get_topic_from_message_info(message)
-                if operand.lower() != topic_name.lower():
-                    return False
-            elif operator == "sender":
-                if operand.lower() != message["sender_email"].lower():
-                    return False
-            elif operator == "is" and operand in ["dm", "private"]:
-                # "is:private" is a legacy alias for "is:dm"
-                if message["type"] != "private":
-                    return False
-            elif operator == "is" and operand in ["starred"]:
-                if operand not in flags:
-                    return False
-            elif operator == "is" and operand == "unread":
-                if "read" in flags:
-                    return False
-            elif operator == "is" and operand in ["alerted", "mentioned"]:
-                if "mentioned" not in flags:
-                    return False
-            elif operator == "is" and operand == "resolved":
-                if message["type"] != "stream":
-                    return False
-                topic_name = get_topic_from_message_info(message)
-                if not topic_name.startswith(RESOLVED_TOPIC_PREFIX):
-                    return False
-            return True
-
-        for narrow_term in narrow:
-            # TODO: Eventually handle negated narrow terms.
-            if not satisfies_operator(operator=narrow_term.operator, operand=narrow_term.operand):
-                return False
-
-        return True
-
-    return narrow_predicate
 
 
 LARGER_THAN_MAX_MESSAGE_ID = 10000000000000000
